Rename contexts and adjust structure of providers
