+++ ui/app/results/ResultGrid.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { DataGrid, GridColDef, GridRowsProp, GridToolbar } from '@mui/x-data-grid';
 import { Paper, Switch, FormControlLabel, Box } from '@mui/material';
 import InstructionBox from './InstructionBox';
-import { useData } from '../context/DataProvider';
-import { useError } from '../context/ErrorProvider';
+import { useData } from '../context/DataContext';
+import { useError } from '../context/ErrorContext';
 import { useRouter } from 'next/navigation';
 import { searchPapers } from '../actions';
 import PaperDetail from './PaperDetail';
